Move to a separate file. Add welcome message update.
