+++ lib/twitter/rest/direct_messages.rb
@@ -215,47 +215,6 @@ module Twitter
         Twitter::DirectMessageEvent.new(response[:event])
       end
 
-      def create_welcome_message(text, name: nil)
-        options = {
-                welcome_message: {
-                  message_data: {
-                    text: text
-                  }
-                }
-              }
-        options[:welcome_message][:name] = name if name
-        response = Twitter::REST::Request.new(self, :json_post, '/1.1/direct_messages/welcome_messages/new.json', options).perform
-        response
-      end
-
-      def welcome_message_list
-        response = Twitter::REST::Request.new(self, :get, '/1.1/direct_messages/welcome_messages/list.json').perform
-        response
-      end
-
-      def create_welcome_message_rule(welcome_message_id)
-        options = {
-                welcome_message_rule: {
-                  welcome_message_id: welcome_message_id
-                }
-              }
-        response = Twitter::REST::Request.new(self, :json_post, '/1.1/direct_messages/welcome_messages/rules/new.json', options).perform
-        response
-      end
-
-      def delete_welcome_message_rule(welcome_message_rule_id)
-        options = {
-                id: welcome_message_rule_id
-              }
-        response = Twitter::REST::Request.new(self, :delete, '/1.1/direct_messages/welcome_messages/rules/destroy.json', options).perform
-        response
-      end
-
-      def welcome_message_rule_list
-        response = Twitter::REST::Request.new(self, :get, '/1.1/direct_messages/welcome_messages/rules/list.json').perform
-        response
-      end
-
     private
 
       def format_json_options(user_id, text, options)
