+++ lib/twitter/rest/request.rb
@@ -28,7 +28,7 @@ module Twitter
         set_multipart_options!(request_method, options)
         @path = uri.path
         @options = options
-        @options_key = {get: :params, json_post: :json, delete: :params}[request_method] || :form
+        @options_key = {get: :params, json_post: :json, delete: :params, json_put: :json}[request_method] || :form
       end
 
       # @return [Array, Hash]
@@ -56,6 +56,9 @@ module Twitter
         if %i[multipart_post json_post].include?(request_method)
           merge_multipart_file!(options) if request_method == :multipart_post
           @request_method = :post
+          @headers = Twitter::Headers.new(@client, @request_method, @uri).request_headers
+        elsif %i[json_put].include?(request_method)
+          @request_method = :put
           @headers = Twitter::Headers.new(@client, @request_method, @uri).request_headers
         else
           @request_method = request_method
